- taskDto added - task converter dto -> entity, entity -> dto - KanbanBoard added - add-task method added - form for add task added - delete, update added - addform and archive .html files added - layout added

+++ src/main/java/io/webskyteam/turbokanban/repository/TaskRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import io.webskyteam.turbokanban.entity.TaskEntity;
 
 @Repository
-public interface TaskRepository extends JpaRepository <TaskEntity,Long>{
+public interface TaskRepository extends JpaRepository <TaskEntity,Integer>{
 
 }
